make agent state controllers injectable
