+++ src/Automata/LLM/Agent.php
@@ -28,6 +28,7 @@ use BlueFission\Automata\LLM\Agent\State\AgentModuleResult;
 use BlueFission\Automata\LLM\Agent\State\AgentState;
 use BlueFission\Automata\LLM\Agent\State\CognitiveController;
 use BlueFission\Automata\LLM\Agent\State\IAgentModule;
+use BlueFission\Automata\LLM\Agent\State\IStateController;
 use BlueFission\Automata\Memory\IWorkingMemory;
 use BlueFission\Automata\LLM\MCP\MCPClient;
 use BlueFission\Automata\LLM\MCP\Tools\MCPDiscoveryTool;
@@ -62,7 +63,7 @@ class Agent implements IDispatcher
     protected ?Orchestrator $orchestrator = null;
     protected AgentSession $session;
     protected AgentState $agentState;
-    protected CognitiveController $cognitiveController;
+    protected IStateController $cognitiveController;
     protected ?TaskCallMonitor $callMonitor = null;
     protected ?HumanReviewGate $humanReviewGate = null;
 
@@ -478,7 +479,7 @@ class Agent implements IDispatcher
     /**
      * Replace the high-level cognitive controller.
      */
-    public function setCognitiveController(CognitiveController $controller): void
+    public function setCognitiveController(IStateController $controller): void
     {
         $this->cognitiveController = $controller;
     }
